mvvm ディレクトリ構築アルファ
