--- app/src/main/java/com/example/a774ist/model/entity/Schedule.kt
@@ -1,4 +1,0 @@
-package com.example.a774ist.model.entity
-
-class Schedule {
-}
