+++ app/src/main/java/com/example/android_774ist/service/model/ScheduleResult.kt
@@ -1,0 +1,4 @@
+package com.example.android_774ist.service.model
+
+class ScheduleResult {
+}
