+++ app/src/main/java/com/example/android_774ist/service/repository/scheduleRepository.kt
@@ -1,4 +1,4 @@
-package com.example.a774ist.model.repository
+package com.example.android_774ist.service.repository
 
 class scheduleRepository {
     //APIから値を受け取る処理
