+++ app/src/main/java/com/example/android_774ist/view/ui/DashboardFragment.kt
@@ -1,4 +1,4 @@
-package com.example.a774ist.ui.dashboard
+package com.example.android_774ist.view.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,7 +8,8 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import com.example.a774ist.R
+import com.example.android_774ist.R
+import com.example.android_774ist.viewmodel.DashboardViewModel
 
 class DashboardFragment : Fragment() {
