+++ app/src/main/java/com/example/android_774ist/view/ui/HomeFragment.kt
@@ -1,14 +1,14 @@
-package com.example.a774ist.ui.home
+package com.example.android_774ist.view.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import com.example.a774ist.R
+import com.example.android_774ist.R
+import com.example.android_774ist.viewmodel.HomeViewModel
 
 class HomeFragment : Fragment() {
