+++ app/src/main/java/com/example/android_774ist/viewmodel/NotificationsViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.a774ist.ui.dashboard
+package com.example.android_774ist.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class NotificationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
