+++ app/src/main/res/layout/fragment_home.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.MainScheduleFragment">
+    tools:context=".view.ui.MainScheduleFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
