+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.example.a774ist.ui.home.HomeFragment"
+        android:name="com.example.android_774ist.view.ui.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.example.a774ist.ui.dashboard.DashboardFragment"
+        android:name="com.example.android_774ist.view.ui.DashboardFragment"
         android:label="@string/title_dashboard"
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.a774ist.ui.notifications.NotificationsFragment"
+        android:name="com.example.android_774ist.view.ui.NotificationsFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
 </navigation>
